reorganized project file structure
