+++ src/components/BackToTop/styles/BackToTop.css
@@ -1,0 +1,18 @@
+/* Skip Link Styling */
+.skip-link {
+  position: absolute;
+  left: -9999px;
+  z-index: 999;
+  padding: 1em;
+  background-color: var(--color-background);
+  color: var(--color-text-primary);
+  font-weight: 500;
+  text-decoration: none;
+  transition: left 0s;
+  border-radius: 8px;
+}
+
+.skip-link:focus {
+  left: 50%;
+  transform: translateX(-50%);
+}
