+++ src/pages/Contact/index.jsx
@@ -3,7 +3,7 @@ import { motion } from 'framer-motion';
 import { MetaTags } from './components/MetaTags';
 import { ContactForm } from './components/ContactForm';
 import { StatusMessage } from './components/StatusMessage';
-import { BackToTop } from '../../components/BackToTop';
+import { BackToTop } from '../../components/BackToTop/BackToTop';
 import { useAnimations } from './hooks/useAnimations';
 import { useForm } from './hooks/useForm';
 import './styles/Contact.css';
