+++ src/pages/Experience/index.jsx
@@ -2,7 +2,7 @@ import { MetaTags } from './components/MetaTags';
 import { Skills } from './components/Skills/Skills';
 import { Certificates } from './components/Certificates/Certificates';
 import { Resume } from './components/Resume/Resume';
-import { BackToTop } from '../../components/BackToTop';
+import { BackToTop } from '../../components/BackToTop/BackToTop';
 import './styles/Experience.css';
 
 const Experience = () => {
