+++ src/pages/Projects/index.jsx
@@ -3,7 +3,7 @@ import { motion } from 'framer-motion';
 import { ProjectGrid } from './components/ProjectGrid';
 import { ProjectDetailsModal } from './components/ProjectDetailsModal/ProjectDetailsModal';
 import { MetaTags } from './components/MetaTags';
-import { BackToTop } from '../../components/BackToTop';
+import { BackToTop } from '../../components/BackToTop/BackToTop';
 import { useAnimations } from './hooks/useAnimations';
 import './styles/Projects.css';
